chore: Refactor generator provider architecture.
